@@ -11,9 +11,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4"
     },
     "autoload": {
-        "psr-4": {"Nggit\\PHPTerbilang\\": "src/"}
+        "psr-4": {
+            "Nggit\\PHPTerbilang\\": "src/"
+        }
     }
 }
